Add remote_kernels plugin

+++ plugins/image_viewer/txl_image_viewer/components.py
@@ -3,9 +3,10 @@ import tempfile
 
 from asphalt.core import Component, Context
 from PIL import Image
-from textual_imageview.viewer import ImageViewer
 from textual.widget import Widget
-from txl.base import Editor, Editors, Contents, FileOpenEvent
+from textual_imageview.viewer import ImageViewer
+
+from txl.base import Contents, Editor, Editors, FileOpenEvent
 from txl.hooks import register_component
 
 
@@ -21,24 +22,35 @@ class _ImageViewer(Editor, Widget, metaclass=ImageViewerMeta):
     def __init__(self, contents: Contents) -> None:
         super().__init__()
         self.contents = contents
+        self.image_viewer = None
 
     async def on_open(self, event: FileOpenEvent) -> None:
         await self.open(event.path)
 
     async def open(self, path: str) -> None:
-        data = await self.contents.get(path, type="bytes", on_change=self.on_change)
+        self.ydoc = await self.contents.get(path, type="blob")
+        self.data = self.ydoc.source
+        self.update_viewer()
+        self.ydoc.observe(self.on_change)
+
+    def update_viewer(self):
+        if not self.data:
+            return
         f = tempfile.NamedTemporaryFile(delete=False)
         try:
-            f.write(data)
+            f.write(self.data)
             f.close()
         finally:
             image = Image.open(f.name)
             os.unlink(f.name)
-        image_viewer = ImageViewer(image)
-        self.mount(image_viewer)
+        if self.image_viewer:
+            self.image_viewer.remove()
+        self.image_viewer = ImageViewer(image)
+        self.mount(self.image_viewer)
 
-    def on_change(self, value):
-        pass
+    def on_change(self, target, event):
+        self.data = self.ydoc.source
+        self.update_viewer()
 
 
 class ImageViewerComponent(Component):
@@ -57,7 +69,9 @@ class ImageViewerComponent(Component):
 
         if self.register:
             editors = await ctx.request_resource(Editors, "editors")
-            editors.register_editor_factory(image_viewer_factory, [".png"])
+            editors.register_editor_factory(
+                image_viewer_factory, [".png", ".jpg", ".jpeg"]
+            )
         else:
             image_viewer = image_viewer_factory()
             ctx.add_resource(image_viewer, name="image_viewer", types=Editor)
